여러가지 수정 entity 각 PK값들 데이터 타입을 Long으로 변경하였습니다 폴더 위치 수정하였습니다.

+++ src/main/java/com/green/muziuniv_be_notuser/entity/schedule/Schedule.java
@@ -19,7 +19,7 @@ public class Schedule extends CreatedAt {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int scheduleId;
+    private Long scheduleId;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "semester_id", nullable = false)
